refactor: extract useExportImportState, useLogsState, useMiscUiState
Three more leaf hooks extracted from AppContext:
- useExportImportState: 10 state vars + 2 refs + 2 callbacks (150 lines)
- useLogsState: 6 state vars + loadLogs callback (52 lines)
- useMiscUiState: 23 state vars for MCP/games/UI chrome (137 lines)

AppContext.tsx: 6,992 → 6,904 lines (-88)
Cumulative: 8,366 → 6,904 (-1,462 lines, 10 hooks extracted)
